chore(tests): reorganize by splitting one test per file in tasks.
