+++ tests/tasks/econvention_to_oscar/test_transform_econvention_to_oscar.py
@@ -1,62 +1,16 @@
-import datetime
+from datetime import datetime
 import json
-import time
 import pytest
 from airflow.utils.state import TaskInstanceState
 from models.econvention_factory import EconventionFactory
 from tests.utils.dag import (
-    assert_dag_dict_equal,
     DATA_INTERVAL_START,
     DATA_INTERVAL_END,
     create_dag_run,
     create_task_instance,
 )
 
-TRANSFORM_TASK_NAME = "dags.tasks.transform" + ".transform_from_econvention_to_oscar"
-
-
-def test_dag_loaded(econvention_to_oscar_dag):
-    """
-    Test dag loading
-    """
-    assert econvention_to_oscar_dag is not None
-    assert len(econvention_to_oscar_dag.tasks) == 4
-
-
-def test_dag_structure(econvention_to_oscar_dag):
-    """
-    Test that the DAG has the correct structure
-    """
-    expected_structure = {
-        "extract_from_econvention": ["transform_from_econvention_to_oscar"],
-        "transform_from_econvention_to_oscar": ["load"],
-        "load": ["load_to_oscar"],
-        "load_to_oscar": [],  # END
-    }
-    assert_dag_dict_equal(expected_structure, econvention_to_oscar_dag)
-
-
-def test_extract_from_econvention(
-    econvention_to_oscar_dag, econvention_raw_data, unique_logical_date
-):
-    """
-    Test the `extract_from_econvention` function to verify correct parsing
-    and loading of JSON data used in the Airflow DAG pipeline.
-    Note:
-    This test uses static sample data located in `/tests/data/` to ensure repeatability.
-    """
-
-    dag_run = create_dag_run(
-        dag=econvention_to_oscar_dag,
-        data_interval_start=DATA_INTERVAL_START,
-        data_interval_end=DATA_INTERVAL_END,
-        logical_date=unique_logical_date,
-        conf_data={"items": econvention_raw_data},
-    )
-    ti = create_task_instance(
-        econvention_to_oscar_dag, dag_run, "extract_from_econvention"
-    )
-    assert ti.state == TaskInstanceState.SUCCESS
+TRANSFORM_TASK_NAME = "dags.tasks.transform.transform_from_econvention_to_oscar"
 
 
 @pytest.mark.parametrize(
@@ -108,7 +62,7 @@ def test_transform_econvention_to_oscar(
     explicitly set in the expected data, as the source file is static and does not
     auto-update with the current date.
     """
-    date_iso = str(datetime.date.fromtimestamp(time.time()).isoformat())
+    date_iso = str(datetime.now().date().isoformat())
 
     # update dynamically any attributes that require ISO date
     for item in oscar_expected_data:
